+++ tests/framework/web/JsonResponseFormatterTest.php
@@ -21,9 +21,9 @@ class JsonResponseFormatterTest extends FormatterTest
     /**
      * @return JsonResponseFormatter
      */
-    protected function getFormatterInstance()
+    protected function getFormatterInstance($configuration = [])
     {
-        return new JsonResponseFormatter();
+        return new JsonResponseFormatter($configuration);
     }
 
     public function formatScalarDataProvider()
@@ -106,6 +106,57 @@ class JsonResponseFormatterTest extends FormatterTest
         ];
     }
 
+    public function contentTypeGenerationDataProvider()
+    {
+        return [
+            [
+                [
+                ],
+                'application/json; charset=UTF-8',
+            ],
+            [
+                [
+                    'useJsonp' => false,
+                ],
+                'application/json; charset=UTF-8',
+            ],
+            [
+                [
+                    'useJsonp' => true,
+                ],
+                'application/javascript; charset=UTF-8',
+            ],
+            [
+                [
+                    'contentType' => 'application/javascript; charset=UTF-8',
+                    'useJsonp' => false,
+                ],
+                'application/javascript; charset=UTF-8',
+            ],
+            [
+                [
+                    'contentType' => 'application/json; charset=UTF-8',
+                    'useJsonp' => true,
+                ],
+                'application/json; charset=UTF-8',
+            ],
+            [
+                [
+                    'contentType' => 'application/hal+json; charset=UTF-8',
+                    'useJsonp' => false,
+                ],
+                'application/hal+json; charset=UTF-8',
+            ],
+            [
+                [
+                    'contentType' => 'application/hal+json; charset=UTF-8',
+                    'useJsonp' => true,
+                ],
+                'application/hal+json; charset=UTF-8',
+            ],
+        ];
+    }
+
     /**
      * @param mixed  $data the data to be formatted
      * @param string $json the expected JSON body
@@ -118,5 +169,19 @@ class JsonResponseFormatterTest extends FormatterTest
         $this->formatter->prettyPrint = true;
         $this->formatter->format($this->response);
         $this->assertEquals($prettyJson, $this->response->content);
+    }
+
+    /**
+     * @param array $configuration JSON formatter configuration array.
+     * @param string $contentTypeExpected Expected value of the response `Content-Type` header.
+     * @dataProvider contentTypeGenerationDataProvider
+     */
+    public function testContentTypeGeneration($configuration, $contentTypeExpected)
+    {
+        $formatter = $this->getFormatterInstance($configuration);
+        $formatter->format($this->response);
+        $contentTypeActual = $this->response->headers->get('Content-Type');
+
+        $this->assertEquals($contentTypeExpected, $contentTypeActual);
     }
 }