+++ framework/CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 Change Log
 
 - Bug #14276: Fixed I18N format with dotted parameters (developeruz)
 - Enh #15135: Automatic completion for help in bash and zsh (Valkeru)
+- Enh #14662: Added support for custom `Content-Type` specification to `yii\web\JsonResponseFormatter` (Kolyunya)
 
 
 2.0.13.1 November 14, 2017